Semi-Functional Login Test
- Grabbing HTML values from POST request

- Basic HTML for Login Page (will make prettier I promise)

- Experimenting with Login Manager
