+++ backend/AdminPortal/login.html
@@ -1,17 +1,14 @@
 <!DOCTYPE html>
 <html>
-<head>
-    <title>Login Test</title>
-</head>
 <body>
     <center>
-     <form method="GET">
-        <p>Username</p>
-        <p><input type="text" name = "username"/></p>
-        <p>Password</p>
-        <p><input type="password" name="password"/></p>
+     <form method="POST">
+        <p>Username:</p>
+        <p><input type="text" name = "user"/></p>
+        <p>Password:</p>
+        <p><input type="password" name="passwrd"/></p>
 
-        <p><input type="submit" value="submit"/></p>
+        <p><input type="submit" value="Sign In"/></p>
      </form>
     </center>
 </body>
